requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqladmin>=0.11.0
sqlalchemy>=2.0.0
alembic>=1.10.0
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.18
pydantic-settings>=2.0.0
email-validator>=2.0.0
requests>=2.32.4
mysql-connector-python>=9.1.0
pymysql>=1.1.1